Add a few helper functions
